Add flag to use the new PR assignment workflow
If the env variable `USE_NEW_PR_ASSIGNMENT` is found, the new PR
assignment will be used, else the old one will still be used. This is
meant as a quick way to test the new workflow and revert to the old one
in case something not working as expected.